requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
email-validator==1.3.1
python-dotenv==1.0.0
